+++ packages/spacecat-shared-rum-api-client/src/common/traffic.js
@@ -26,9 +26,17 @@ import URI from 'urijs';
  */
 export function getSecondLevelDomain(url) {
   if (!hasText(url)) return url;
-  const uri = new URI(prependSchema(url));
-  const tld = uri.tld();
-  return uri.hostname().split(`.${tld}`)[0];
+  if (url === '(direct)') return '';
+
+  try {
+    const uri = new URI(prependSchema(url));
+    const tld = uri.tld();
+    return uri.hostname().split(`.${tld}`)[0];
+    /* c8 ignore next 4 */
+  } catch (error) {
+    // future-proof for the cases where url cannot be parsed for some reason
+    return url;
+  }
 }
 
 /*